@@ -1,30 +1,61 @@
 [project]
 name = "click"
+version = "8.2.1"
 description = "Composable command line interface toolkit"
 readme = "README.md"
-license = {file = "LICENSE.txt"}
+license = "BSD-3-Clause"
+license-files = ["LICENSE.txt"]
 maintainers = [{name = "Pallets", email = "contact@palletsprojects.com"}]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
     "Intended Audience :: Developers",
-    "License :: OSI Approved :: BSD License",
     "Operating System :: OS Independent",
     "Programming Language :: Python",
     "Typing :: Typed",
 ]
-requires-python = ">=3.7"
+requires-python = ">=3.10"
 dependencies = [
     "colorama; platform_system == 'Windows'",
-    "importlib-metadata; python_version < '3.8'",
 ]
-dynamic = ["version"]
 
 [project.urls]
 Donate = "https://palletsprojects.com/donate"
 Documentation = "https://click.palletsprojects.com/"
-Changes = "https://click.palletsprojects.com/changes/"
+Changes = "https://click.palletsprojects.com/page/changes/"
 Source = "https://github.com/pallets/click/"
 Chat = "https://discord.gg/pallets"
+
+[dependency-groups]
+dev = [
+    "ruff",
+    "tox",
+    "tox-uv",
+]
+docs = [
+    "myst-parser",
+    "pallets-sphinx-themes",
+    "sphinx",
+    "sphinx-tabs",
+    "sphinxcontrib-log-cabinet",
+]
+docs-auto = [
+    "sphinx-autobuild",
+]
+gha-update = [
+    "gha-update ; python_full_version >= '3.12'",
+]
+pre-commit = [
+    "pre-commit",
+    "pre-commit-uv",
+]
+tests = [
+    "pytest",
+]
+typing = [
+    "mypy",
+    "pyright",
+    "pytest",
+]
 
 [build-system]
 requires = ["flit_core<4"]
