@@ -1,22 +1,24 @@
+from __future__ import annotations
+
+import collections.abc as cabc
 import os
 import re
 import typing as t
 from gettext import gettext as _
 
 from .core import Argument
-from .core import BaseCommand
+from .core import Command
 from .core import Context
-from .core import MultiCommand
+from .core import Group
 from .core import Option
 from .core import Parameter
 from .core import ParameterSource
-from .parser import split_arg_string
 from .utils import echo
 
 
 def shell_complete(
-    cli: BaseCommand,
-    ctx_args: t.MutableMapping[str, t.Any],
+    cli: Command,
+    ctx_args: cabc.MutableMapping[str, t.Any],
     prog_name: str,
     complete_var: str,
     instruction: str,
@@ -77,12 +79,12 @@ class CompletionItem:
         self,
         value: t.Any,
         type: str = "plain",
-        help: t.Optional[str] = None,
+        help: str | None = None,
         **kwargs: t.Any,
     ) -> None:
         self.value: t.Any = value
         self.type: str = type
-        self.help: t.Optional[str] = help
+        self.help: str | None = help
         self._info = kwargs
 
     def __getattr__(self, name: str) -> t.Any:
@@ -215,8 +217,8 @@ class ShellComplete:
 
     def __init__(
         self,
-        cli: BaseCommand,
-        ctx_args: t.MutableMapping[str, t.Any],
+        cli: Command,
+        ctx_args: cabc.MutableMapping[str, t.Any],
         prog_name: str,
         complete_var: str,
     ) -> None:
@@ -233,7 +235,7 @@ class ShellComplete:
         safe_name = re.sub(r"\W*", "", self.prog_name.replace("-", "_"), flags=re.ASCII)
         return f"_{safe_name}_completion"
 
-    def source_vars(self) -> t.Dict[str, t.Any]:
+    def source_vars(self) -> dict[str, t.Any]:
         """Vars for formatting :attr:`source_template`.
 
         By default this provides ``complete_func``, ``complete_var``,
@@ -253,16 +255,14 @@ class ShellComplete:
         """
         return self.source_template % self.source_vars()
 
-    def get_completion_args(self) -> t.Tuple[t.List[str], str]:
+    def get_completion_args(self) -> tuple[list[str], str]:
         """Use the env vars defined by the shell script to return a
         tuple of ``args, incomplete``. This must be implemented by
         subclasses.
         """
         raise NotImplementedError
 
-    def get_completions(
-        self, args: t.List[str], incomplete: str
-    ) -> t.List[CompletionItem]:
+    def get_completions(self, args: list[str], incomplete: str) -> list[CompletionItem]:
         """Determine the context and last complete command or parameter
         from the complete args. Call that object's ``shell_complete``
         method to get the completions for the incomplete value.
@@ -338,7 +338,7 @@ class BashComplete(ShellComplete):
         self._check_version()
         return super().source()
 
-    def get_completion_args(self) -> t.Tuple[t.List[str], str]:
+    def get_completion_args(self) -> tuple[list[str], str]:
         cwords = split_arg_string(os.environ["COMP_WORDS"])
         cword = int(os.environ["COMP_CWORD"])
         args = cwords[1:cword]
@@ -360,7 +360,7 @@ class ZshComplete(ShellComplete):
     name = "zsh"
     source_template = _SOURCE_ZSH
 
-    def get_completion_args(self) -> t.Tuple[t.List[str], str]:
+    def get_completion_args(self) -> tuple[list[str], str]:
         cwords = split_arg_string(os.environ["COMP_WORDS"])
         cword = int(os.environ["COMP_CWORD"])
         args = cwords[1:cword]
@@ -382,7 +382,7 @@ class FishComplete(ShellComplete):
     name = "fish"
     source_template = _SOURCE_FISH
 
-    def get_completion_args(self) -> t.Tuple[t.List[str], str]:
+    def get_completion_args(self) -> tuple[list[str], str]:
         cwords = split_arg_string(os.environ["COMP_WORDS"])
         incomplete = os.environ["COMP_CWORD"]
         args = cwords[1:]
@@ -401,10 +401,10 @@ class FishComplete(ShellComplete):
         return f"{item.type},{item.value}"
 
 
-ShellCompleteType = t.TypeVar("ShellCompleteType", bound=t.Type[ShellComplete])
+ShellCompleteType = t.TypeVar("ShellCompleteType", bound="type[ShellComplete]")
 
 
-_available_shells: t.Dict[str, t.Type[ShellComplete]] = {
+_available_shells: dict[str, type[ShellComplete]] = {
     "bash": BashComplete,
     "fish": FishComplete,
     "zsh": ZshComplete,
@@ -412,7 +412,7 @@ _available_shells: t.Dict[str, t.Type[ShellComplete]] = {
 
 
 def add_completion_class(
-    cls: ShellCompleteType, name: t.Optional[str] = None
+    cls: ShellCompleteType, name: str | None = None
 ) -> ShellCompleteType:
     """Register a :class:`ShellComplete` subclass under the given name.
     The name will be provided by the completion instruction environment
@@ -431,7 +431,7 @@ def add_completion_class(
     return cls
 
 
-def get_completion_class(shell: str) -> t.Optional[t.Type[ShellComplete]]:
+def get_completion_class(shell: str) -> type[ShellComplete] | None:
     """Look up a registered :class:`ShellComplete` subclass by the name
     provided by the completion instruction environment variable. If the
     name isn't registered, returns ``None``.
@@ -439,6 +439,43 @@ def get_completion_class(shell: str) -> t.Optional[t.Type[ShellComplete]]:
     :param shell: Name the class is registered under.
     """
     return _available_shells.get(shell)
+
+
+def split_arg_string(string: str) -> list[str]:
+    """Split an argument string as with :func:`shlex.split`, but don't
+    fail if the string is incomplete. Ignores a missing closing quote or
+    incomplete escape sequence and uses the partial token as-is.
+
+    .. code-block:: python
+
+        split_arg_string("example 'my file")
+        ["example", "my file"]
+
+        split_arg_string("example my\\")
+        ["example", "my"]
+
+    :param string: String to split.
+
+    .. versionchanged:: 8.2
+        Moved to ``shell_completion`` from ``parser``.
+    """
+    import shlex
+
+    lex = shlex.shlex(string, posix=True)
+    lex.whitespace_split = True
+    lex.commenters = ""
+    out = []
+
+    try:
+        for token in lex:
+            out.append(token)
+    except ValueError:
+        # Raised when end-of-string is reached in an invalid state. Use
+        # the partial token as-is. The quote or escape character is in
+        # lex.state, not lex.token.
+        out.append(lex.token)
+
+    return out
 
 
 def _is_incomplete_argument(ctx: Context, param: Parameter) -> bool:
@@ -475,7 +512,7 @@ def _start_of_option(ctx: Context, value: str) -> bool:
     return c in ctx._opt_prefixes
 
 
-def _is_incomplete_option(ctx: Context, args: t.List[str], param: Parameter) -> bool:
+def _is_incomplete_option(ctx: Context, args: list[str], param: Parameter) -> bool:
     """Determine if the given parameter is an option that needs a value.
 
     :param args: List of complete args before the incomplete value.
@@ -500,10 +537,10 @@ def _is_incomplete_option(ctx: Context, args: t.List[str], param: Parameter) -> 
 
 
 def _resolve_context(
-    cli: BaseCommand,
-    ctx_args: t.MutableMapping[str, t.Any],
+    cli: Command,
+    ctx_args: cabc.MutableMapping[str, t.Any],
     prog_name: str,
-    args: t.List[str],
+    args: list[str],
 ) -> Context:
     """Produce the context hierarchy starting with the command and
     traversing the complete arguments. This only follows the commands,
@@ -514,51 +551,55 @@ def _resolve_context(
     :param args: List of complete args before the incomplete value.
     """
     ctx_args["resilient_parsing"] = True
-    ctx = cli.make_context(prog_name, args.copy(), **ctx_args)
-    args = ctx.protected_args + ctx.args
+    with cli.make_context(prog_name, args.copy(), **ctx_args) as ctx:
+        args = ctx._protected_args + ctx.args
 
-    while args:
-        command = ctx.command
+        while args:
+            command = ctx.command
 
-        if isinstance(command, MultiCommand):
-            if not command.chain:
-                name, cmd, args = command.resolve_command(ctx, args)
-
-                if cmd is None:
-                    return ctx
-
-                ctx = cmd.make_context(name, args, parent=ctx, resilient_parsing=True)
-                args = ctx.protected_args + ctx.args
-            else:
-                sub_ctx = ctx
-
-                while args:
+            if isinstance(command, Group):
+                if not command.chain:
                     name, cmd, args = command.resolve_command(ctx, args)
 
                     if cmd is None:
                         return ctx
 
-                    sub_ctx = cmd.make_context(
-                        name,
-                        args,
-                        parent=ctx,
-                        allow_extra_args=True,
-                        allow_interspersed_args=False,
-                        resilient_parsing=True,
-                    )
-                    args = sub_ctx.args
+                    with cmd.make_context(
+                        name, args, parent=ctx, resilient_parsing=True
+                    ) as sub_ctx:
+                        ctx = sub_ctx
+                        args = ctx._protected_args + ctx.args
+                else:
+                    sub_ctx = ctx
 
-                ctx = sub_ctx
-                args = [*sub_ctx.protected_args, *sub_ctx.args]
-        else:
-            break
+                    while args:
+                        name, cmd, args = command.resolve_command(ctx, args)
+
+                        if cmd is None:
+                            return ctx
+
+                        with cmd.make_context(
+                            name,
+                            args,
+                            parent=ctx,
+                            allow_extra_args=True,
+                            allow_interspersed_args=False,
+                            resilient_parsing=True,
+                        ) as sub_sub_ctx:
+                            sub_ctx = sub_sub_ctx
+                            args = sub_ctx.args
+
+                    ctx = sub_ctx
+                    args = [*sub_ctx._protected_args, *sub_ctx.args]
+            else:
+                break
 
     return ctx
 
 
 def _resolve_incomplete(
-    ctx: Context, args: t.List[str], incomplete: str
-) -> t.Tuple[t.Union[BaseCommand, Parameter], str]:
+    ctx: Context, args: list[str], incomplete: str
+) -> tuple[Command | Parameter, str]:
     """Find the Click object that will handle the completion of the
     incomplete value. Return the object and the incomplete value.
 
