@@ -1,4 +1,4 @@
-"""From https://click.palletsprojects.com/en/8.1.x/quickstart/#adding-parameters"""
+"""From https://click.palletsprojects.com/en/stable/quickstart/#adding-parameters"""
 
 from typing_extensions import assert_type
 
